JP-147: contain the canvas toolbar in the canvas region
Split the single app-wide UnifiedToolbar into a slim global app bar and a
region-scoped CanvasToolbar so canvas controls stop leaking across the app.

- New CanvasToolbar (rendered inside .canvas-area-wrapper): drawing tools,
  shape/custom pickers, file import, rebuild-connectors, undo/redo, and the
  canvas page tabs. Because it lives in the canvas region it auto-hides with
  the canvas in Relaxed `write` focus (no leak) and `flex-wrap`s downward in
  the narrow Relaxed `split` pane instead of overflowing horizontally.
- Extract InlinePageTabs into its own file.
- UnifiedToolbar slims to app chrome only: document name + save status, the
  Relaxed focus switch, layout selector, whiteboard, help, settings. Dropped
  the now-moved getImportContext / onRebuildConnectors props.
- App.tsx renders <CanvasToolbar> above <CanvasContainer>; .canvas-area-wrapper
  becomes a column (toolbar stacks above the canvas; .canvas-area min-height:0).

Reuses the existing .tool-button / .toolbar-* / .inline-* styles. Second slice
of the JP-147 editor visual-design pass. typecheck clean, 1734 tests pass,
build green.
